backend/fastapi_consumer/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
gunicorn==20.1.0
orjson==3.8.4
aiokafka==0.8.0
PyJWT==2.6.0